+++ Assets/Scripts/InGame/Player/PlayerStats.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class PlayerStats : ScriptableObject
 {
     public float baseMoveSpeed = 3.5f;
-    public int baseMaxHP = 100;
+    public float baseMaxHP = 100;
     public float baseAttackPower = 10f;
     public float baseInteractRange = 1.2f;
 }
